feat: Overhaul RL training system for better convergence - simulation_enhanced.py: Complete rewrite with proper reward shaping   - Reduced reward variance (+1.5 to +2.3 for corrections, -0.3 for ignored)   - Added time bonus for quick corrections   - Realistic user behavior simulation (fatigue, compliance, motivation)   - 18-dim state space with user profile features   - Auto-saves JSON results with timestamp - config.py: RL stability improvements   - batch_size: 64→128, memory_size: 10k→20k   - Added lr_decay (0.5) and lr_decay_interval (500)   - learning_rate: starts at 1e-3, decays over training - rl_ppo_agent.py: Added LR scheduler, ppo_epochs 10→5 - rl_agent.py: Added LR scheduler with same decay - main.py: Cleanup   - Removed duplicate enable_dashboard/use_enhanced_training lines   - Updated imports to use simulation_enhanced only   - Added --difficulty option (easy/medium/hard) - environment.py: Removed duplicate 'return reward' statement - visualize_results.py: New training visualization tool   - Auto-detects latest JSON files   - Generates ppo/dqn training curves, comparison plots, benchmark metrics   - Supports --benchmark flag for correction rate analysis - docs: Updated README.md and USAGE_GUIDE.md - cleanup: Deleted simulation.py (replaced by simulation_enhanced.py)

+++ utils.py
@@ -112,18 +112,18 @@ class SessionLogger:
 
     def get_summary(self) -> Dict[str, Any]:
         summary = {
-            "session_name": self.data["session_name"], "total_runtime": self.data["total_runtime"],
+            "session_name": self.data["session_name"], "total_runtime": float(self.data["total_runtime"]),
             "total_frames": len(self.data["posture_scores"]),
             "alerts_sent": self.data["alerts_sent"], "corrections_made": self.data["corrections_made"],
         }
         if self.data["posture_scores"]:
             scores = self.data["posture_scores"]
-            summary["avg_posture_score"] = np.mean(scores)
-            summary["max_posture_score"] = np.max(scores)
-            summary["min_posture_score"] = np.min(scores)
+            summary["avg_posture_score"] = float(np.mean(scores))
+            summary["max_posture_score"] = float(np.max(scores))
+            summary["min_posture_score"] = float(np.min(scores))
         if self.data["rewards"]:
-            summary["total_reward"] = np.sum(self.data["rewards"])
-            summary["avg_reward"] = np.mean(self.data["rewards"])
+            summary["total_reward"] = float(np.sum(self.data["rewards"]))
+            summary["avg_reward"] = float(np.mean(self.data["rewards"]))
         return summary
